requirements.txt
fastapi
uvicorn
pydantic
# sentence-transformers
openai
python-dotenv
gunicorn